delete generator folder
