+++ bridges/modules/grandpa/src/lib.rs
@@ -844,7 +844,6 @@ mod tests {
 	use frame_support::{
 		assert_err, assert_noop, assert_ok,
 		dispatch::{Pays, PostDispatchInfo},
-		storage::generator::StorageValue,
 	};
 	use frame_system::{EventRecord, Phase};
 	use sp_consensus_grandpa::{ConsensusLog, GRANDPA_ENGINE_ID};
