+++ bridges/modules/messages/src/tests/pallet_tests.rs
@@ -40,7 +40,6 @@ use codec::Encode;
 use frame_support::{
 	assert_noop, assert_ok,
 	dispatch::Pays,
-	storage::generator::{StorageMap, StorageValue},
 	traits::Hooks,
 	weights::Weight,
 };
