+++ substrate/frame/support/src/storage/types/counted_map.rs
@@ -19,21 +19,22 @@
 
 use crate::{
 	storage::{
-		generator::StorageMap as _,
 		types::{
 			OptionQuery, QueryKindTrait, StorageEntryMetadataBuilder, StorageMap, StorageValue,
 			ValueQuery,
 		},
-		StorageAppend, StorageDecodeLength, StorageTryAppend,
+		StorageAppend,
+		StorageDecodeLength,
+		StorageTryAppend,
 	},
 	traits::{Get, GetDefault, StorageInfo, StorageInfoTrait, StorageInstance},
 	Never,
 };
-use alloc::{vec, vec::Vec};
 use codec::{Decode, Encode, EncodeLike, FullCodec, MaxEncodedLen, Ref};
 use sp_io::MultiRemovalResults;
 use sp_metadata_ir::StorageEntryMetadataIR;
 use sp_runtime::traits::Saturating;
+use sp_std::prelude::*;
 
 /// A wrapper around a [`StorageMap`] and a [`StorageValue`] (with the value being `u32`) to keep
 /// track of how many items are in a map, without needing to iterate all the values.
@@ -155,7 +156,6 @@ where
 
 	/// The prefix used to generate the key of the map.
 	pub fn map_storage_final_prefix() -> Vec<u8> {
-		use crate::storage::generator::StorageMap;
 		<Self as MapWrapper>::Map::prefix_hash().to_vec()
 	}
